Impl password reset commands
